Draw the edges of the domain (a square in 2D, a cube in 3D).
Now the edges of the domain are drawn on the graphical
visualisation. Theses edges are useful to see the bounds of the
domain.

The structure changed a little because the previous code was assuming
that the only actors in the renderer were representing boids.
Now, the pointers over boids actors are copied in a separate structure
and the actors of the edges of the domain are added to the actors of
the renderer, nearby the actors of the boids.