Generalisation of sample sort to any data.
The sample sort was already templated but was using MPI_INT to exchange data.
Now, it use MPI_BYTE (warning to endianness) and accept a comparator in
order to sort the data according to this comparator.

WARNING: if the connected computers do not have the same endianness, the algorithm
will probably fail. I need to check how to generalize this.